+++ src/sections/episode-fav-browser.ts
@@ -68,7 +68,7 @@ export class EpisodeFavBrowser extends FavBrowserBase {
             return (
               html`<spc-media-browser-list 
                         class="media-browser-list"
-                        .items=${this.mediaList?.filter((item: IEpisode) => item.name.toLocaleLowerCase().indexOf(filterName) !== -1)}
+                        .items=${this.mediaList?.filter((item: IEpisode) => (item.name.toLocaleLowerCase().indexOf(filterName) !== -1) || (item.show.name.toLocaleLowerCase().indexOf(filterName) !== -1))}
                         .store=${this.store}
                         @item-selected=${this.onItemSelected}
                         @item-selected-with-hold=${this.onItemSelectedWithHold}
@@ -78,7 +78,7 @@ export class EpisodeFavBrowser extends FavBrowserBase {
             return (
               html`<spc-media-browser-icons 
                         class="media-browser-list"
-                        .items=${this.mediaList?.filter((item: IEpisode) => item.name.toLocaleLowerCase().indexOf(filterName) !== -1)}
+                        .items=${this.mediaList?.filter((item: IEpisode) => (item.name.toLocaleLowerCase().indexOf(filterName) !== -1) || (item.show.name.toLocaleLowerCase().indexOf(filterName) !== -1))}
                         .store=${this.store}
                         @item-selected=${this.onItemSelected}
                         @item-selected-with-hold=${this.onItemSelectedWithHold}