+++ src/sections/show-fav-browser.ts
@@ -68,7 +68,7 @@ export class ShowFavBrowser extends FavBrowserBase {
                 return (
                   html`<spc-media-browser-list 
                         class="media-browser-list"
-                        .items=${this.mediaList?.filter((item: IShowSimplified) => item.name.toLocaleLowerCase().indexOf(filterName) !== -1)}
+                        .items=${this.mediaList?.filter((item: IShowSimplified) => (item.name.toLocaleLowerCase().indexOf(filterName) !== -1))}
                         .store=${this.store}
                         @item-selected=${this.onItemSelected}
                         @item-selected-with-hold=${this.onItemSelectedWithHold}
@@ -78,7 +78,7 @@ export class ShowFavBrowser extends FavBrowserBase {
                 return (
                   html`<spc-media-browser-icons class="media-browser-list"
                         class="media-browser-list"
-                        .items=${this.mediaList?.filter((item: IShowSimplified) => item.name.toLocaleLowerCase().indexOf(filterName) !== -1)}
+                        .items=${this.mediaList?.filter((item: IShowSimplified) => (item.name.toLocaleLowerCase().indexOf(filterName) !== -1))}
                         .store=${this.store}
                         @item-selected=${this.onItemSelected}
                         @item-selected-with-hold=${this.onItemSelectedWithHold}